+++ changelog.md
@@ -28,6 +28,7 @@ Codion Change Log
 - EntityTablePanel.Settings.includeSummaryPanel() added along with INCLUDE_SUMMARY_PANEL configuration value.
 - EntityTablePanel, column summary field horizontal alignment no follows column cell renderer alignment.
 - EntityTablePanel.Settings.includeEntityMenu() added.
+- EntityTablePanel.INCLUDE_POPUP_MENU configuration value added.
 ### is.codion.swing.framework.server.monitor
 - ServerMonitor and ServerMonitorPanel, tables migrated from plain JTable to FilteredTable.